Move changes to RabbitStreamConfiguration
